bedre null-håndtering
